% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licornetics.R
\name{licornetics}
\alias{licornetics}
\title{Li-COR plotting package}
\usage{
licornetics(
  identifier,
  type = "gsw",
  area_correction = 1,
  timestamps = NULL,
  observations = NULL,
  legend_title = "Genotype",
  legend_labels = identifier,
  remove_outliers = "no"
)
}
\arguments{
\item{identifier}{Keywords that distinguish the Li-COR .xlsx files for the different datasets (e.g. "wt", "mutant1").}

\item{type}{Determines y axis data. Can be "gsw" (absolute stomatal conductance), "relgsw" (relative stomatal conductance), "A" (CO2 assimilation) or "WUE" (water use efficiency).}

\item{area_correction}{Li-COR chamber size divided by average of measured leaf areas. Default is set to 1.}

\item{timestamps}{Optionally add vertical lines to the plot as timeline indicators (e.g. 'c(20, 40, 60)').}

\item{observations}{Crop the range of observations (time) you want to show (e.g. 'c(16:70)').}

\item{legend_title}{Change the title of the legend. Default is set to "Genotype".}

\item{legend_labels}{Change the labels within the legend. Default is set to the input given by the 'identifier' argument.}

\item{remove_outliers}{Optionally remove boxplot outliers (based on data obtained by calculating A/gsw).}
}
\description{
Creates plots for physiological parameters measured using the Li-COR photosystem.
}
\keyword{assimilation}
\keyword{co2}
\keyword{conductance}
\keyword{efficiency}
\keyword{kinetics}
\keyword{leaf}
\keyword{li-cor}
\keyword{photosynthesis}
\keyword{physiology}
\keyword{plot}
\keyword{stomatal}
\keyword{water-use}
\keyword{wue}
