% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licorvalues.R
\name{licorvalues}
\alias{licorvalues}
\title{Calculate steady state and kinetics values from Li-COR data}
\usage{
licorvalues(
  identifier,
  transition = list("All"),
  area_correction = 1,
  stomden = NULL,
  label = NULL,
  remove_outliers = "no",
  colours = NULL,
  errorbars = "se"
)
}
\arguments{
\item{identifier}{Keywords that distinguish the Li-COR .xlsx files for the different datasets (e.g. "wt", "mutant1").}

\item{transition}{Split your data into different time frames to distinguish between transition phases. Default uses all time points at once.}

\item{area_correction}{Correct data by leaf area. Li-COR chamber size divided by average of measured leaf areas. Default is set to 1.}

\item{stomden}{Insert stomatal density to normalise values by stomatal density.}

\item{label}{Change the names for each 'identifier' input. Default is set to the input given by the 'identifier' argument.}

\item{remove_outliers}{Optionally remove boxplot outliers by setting to "yes" (based on outliers from the 'A' column).}

\item{colours}{Set colour to be used for each identifier in final plots (e.g. 'c("red", "green")'). Default is black.}

\item{errorbars}{Set type of errorbars to either standard error ("se", this is the default) or standard deviation ("sd").}
}
\description{
Calculate steady state values and kinetics parameters and plot regression curves from data measured using the Li-COR photosystem.
}
\keyword{assimilation}
\keyword{carbon}
\keyword{co2}
\keyword{conductance}
\keyword{efficiency}
\keyword{exchange}
\keyword{gas}
\keyword{kinetics}
\keyword{leaf}
\keyword{li-cor}
\keyword{photosynthesis}
\keyword{physiology}
\keyword{plot}
\keyword{stomatal}
\keyword{water}
\keyword{water-use}
\keyword{wue}
