% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licorvalues.R
\name{licorvalues}
\alias{licorvalues}
\title{Calculate steady state and kinetics values from Li-COR data.}
\usage{
licorvalues(
  identifier,
  transition = list("All"),
  stomden = NULL,
  label = NULL,
  show_individuals = F,
  remove_outliers = "no",
  colours = NULL,
  errorbars = "se"
)
}
\arguments{
\item{identifier}{Keywords that distinguish the Li-COR .xlsx files for the different datasets (e.g. "wt", "mutant1").}

\item{transition}{Split your data into different time frames to distinguish between transition phases. Default uses all time points at once. Input as lists (e.g. 'list(c(14:25), c(26:37))')}

\item{stomden}{Normalise data by stomatal density (in unit stomata/mm2). Order of input must follow order of identifiers. (e.g. 'c(50, 100, 70)')}

\item{show_individuals}{Change final plot output in R to depict plots for each single input file. Default shows averaged plots for each identifier.}

\item{remove_outliers}{Optionally remove boxplot outliers by setting to "yes" (based on outliers from the 'A' column).}

\item{colours}{Set colour to be used for each identifier in final plots (e.g. 'c("red", "green")'). Default is black.}
}
\description{
Calculate steady state values and kinetics parameters and plot regression curves from data measured using the Li-COR photosystem.
}
\keyword{assimilation}
\keyword{carbon}
\keyword{co2}
\keyword{conductance}
\keyword{efficiency}
\keyword{exchange}
\keyword{gas}
\keyword{kinetics}
\keyword{leaf}
\keyword{li-cor}
\keyword{photosynthesis}
\keyword{physiology}
\keyword{plot}
\keyword{stomatal}
\keyword{water}
\keyword{water-use}
\keyword{wue}
